# Golang/2/rest/docs/swagger.yaml
definitions:
  main.ErrorResponse:
    description: Response for an error
    properties:
      error:
        type: string
    type: object
  main.SuccessResponse:
    description: Response for a successful operation
    properties:
      data:
        items:
          $ref: '#/definitions/main.UserResponse'
        type: array
    type: object
  main.UserResponse:
    description: User response model
    properties:
      age:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /users:
    get:
      description: Get all users with optional filtering, sorting, and pagination
      parameters:
      - description: Age to filter
        in: query
        name: age
        type: integer
      - description: Field to sort by
        in: query
        name: sortBy
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit number of users
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get all users
    post:
      consumes:
      - application/json
      description: Create a new user with validation
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.UserResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Create a new user
  /users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Delete a user by ID
    put:
      consumes:
      - application/json
      description: Update an existing user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.UserResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Update an existing user by ID
swagger: "2.0"